requirements.txt
sanic
sanic-mongodb-extension
pymongo[srv]
sanic-cors
pyjwt
